+++ quartz.layout.ts
@@ -19,6 +19,7 @@ export const defaultContentPageLayout: PageLayout = {
   beforeBody: [
     Component.ArticleTitle(),
     Component.ContentMeta(),
+    Component.FrontmatterDisplay(),
   ],
   left: [
     Component.PageTitle(),